Fix: Project invitation link and config files
