+++ apps/web/frontend/components/projectForm/projectForm.tsx
@@ -13,7 +13,7 @@ import { FragmentType, graphql, useFragment } from '../../generated/gql'
 import { ProjectInput } from '../../generated/gql/graphql'
 import { CalendarSelector } from '../calendarSelector'
 import { PageHeading } from '../pageHeading'
-import { ProjectInvitation } from '../projectInvitation'
+import { ProjectInvitationButton } from '../projectInvitationButton'
 import { DeleteOrArchiveProjectButton } from './deleteOrArchiveProjectButton'
 
 const getDate = (dateString: string | undefined | null): Date | undefined => {
@@ -233,7 +233,8 @@ export const ProjectForm = (props: ProjectFormProps): JSX.Element => {
           </button>
         )}
         {hasError && <span className="display: inline-block pt-5 text-red-600">Unable to save project.</span>}
-        {project && <ProjectInvitation projectId={project.id} />}
+
+        {project?.canModify && project && <ProjectInvitationButton projectId={project.id} />}
       </div>
     </div>
   )
