+++ packages/backend/src/graphql/projectInvitation/projectInvitation.ts
@@ -4,7 +4,7 @@ export const ProjectInvitation = builder.prismaObject('ProjectInvitation', {
   fields: (t) => ({
     id: t.exposeID('id'),
     project: t.relation('project'),
-    expireDate: t.expose('expireDate', { type: 'Date' }),
+    expireDate: t.expose('expireDate', { type: 'DateTime' }),
     invitationKey: t.exposeString('invitationKey'),
   }),
 })
