+++ packages/pulumi/index.ts
@@ -28,7 +28,7 @@ const timebook_database = new digitalocean.DatabaseDb(
     protect: true,
   },
 )
-const db_user = new digitalocean.DatabaseUser(
+const database_user = new digitalocean.DatabaseUser(
   'db',
   {
     clusterId: timebook_cluster.id,
@@ -71,7 +71,7 @@ export const timebook = new digitalocean.App('timebook', {
         production: true,
         clusterName: timebook_cluster.name,
         dbName: timebook_database.name,
-        dbUser: db_user.name,
+        dbUser: database_user.name,
       },
     ],
   },
